updated version
this is going to be the website for a while, it's not 100 percent mobile-friendly quite yet but I will work on that. I just needed to push the updated version because I can't have a plain HTML website anymore. This is pure CSS and HTML. I will update this again in the future but right now I don't see myself adding a substantial update soon. I do plan on changing the entire format but not right now.